+++ src/commands/mc.datapack.delete.ts
@@ -1,20 +1,25 @@
 import { rmdir } from "fs";
 import { Datapack, WorldsProvider } from "../class/WorldsProvider";
 import { window } from "vscode";
+import { notifications } from "../class/NotificationProvider";
 
-export default async (provider:WorldsProvider,pack:Datapack)=>{
+export default async (provider: WorldsProvider, pack: Datapack) => {
     {
+        //Tries deleting the dir
         try {
             await rmdir(pack.data.directory, { recursive: true }, (err) => {
+                // Send error if there is an error
                 if (err) {
-                    window.showErrorMessage(`Couldn't remove the datapack: ${err}`);
+                    notifications.sendErrorMessage(`Couldn't remove the datapack: ${err}`, 'worlds.packDeleteErr');
                 } else {
-                    window.showInformationMessage(`Successfully deleted ${pack.data.name}!`);
+                    // Send noti
+                    notifications.sendInformationMessage(`Successfully deleted ${pack.data.name}!`, 'worlds.packDelete');
                     provider.refresh();
                 }
             });
         } catch (error) {
-            window.showErrorMessage(`Couldn't remove the datapack : ${error}`);
+            // Send error if it fails
+            notifications.sendErrorMessage(`Couldn't remove the datapack: ${error}`, 'worlds.packDeleteErr');
         }
     }
-}
+};