+++ src/extension.ts
@@ -1,8 +1,6 @@
 
-import { join } from 'path';
 import * as vscode from 'vscode';
 import { Datapack, World, WorldsProvider } from './class/WorldsProvider';
-import { mkdir, mkdirSync, rmSync, rmdir, writeFile } from 'fs';
 import mcRefresh from './commands/mc.refresh';
 import mcWorldCreate from './commands/mc.world.create';
 import mcDatapackDelete from './commands/mc.datapack.delete';
@@ -10,9 +8,12 @@ import mcOpen from './commands/mc.open';
 
 export function activate(context: vscode.ExtensionContext) {
 
+	// Starting the worlds provider
 	const provider = new WorldsProvider();
 	vscode.window.registerTreeDataProvider('worlds', provider);
 
+
+	// Registering the commands
 	context.subscriptions.push(
 		vscode.commands.registerCommand('mc.refresh', () => mcRefresh(provider)),
 		vscode.commands.registerCommand('mc.world.create', async (world: World<Datapack>) => mcWorldCreate(provider, world)),
@@ -21,5 +22,4 @@ export function activate(context: vscode.ExtensionContext) {
 	);
 }
 
-// this method is called when your extension is deactivated
 export function deactivate() { }